NewCapDispenser
For BeanWS0.1 this implements a prototype for the new worm gear cap dispenser.

For BeanWS0.2 this deletes everything in the old version and is a copy and paste of the WS0.1 except it has a prototype software for the new cap dispenser when each stepper motor is connected directly to the arduino.