Y Axis Calibration
Fixed the stepper motor problem for the Y Axis Calibration. The gripper was going down too slow and I probably fixed that but have not tested it. The gripper will not stop going down now, it is likely an electrical problem with the switch, but it could be a sw bug.