rough draft calibration
The calibration program is complete however there are things that could be cleaned up and added.
There does seem to be errors in the final location that the x axis goes to. However these errors are not present in the y axis despite nearly identical code which leads me to believe the stepper motor system for the x axis is the problem.