New Steps size
The step sized has been decreased so the locations and speed have been changed for the x stepper motor.

The press base has been lowered so the y location has been changed and the press height needs to be changed.

Do not change the height to 8 for the large cup since this new setting likely broke the press.