60%
This is able to complete the coffee 60% of the time. It must recalibrate between each coffee likely because of the lack of amps. The errors are either due to the stepper motors error or because it is calibrating properly. If it is a calibration issue try having the X axis calibrate the standard way and then have the gripper go forward a short distance and recalibrate slowly.