Finishing Touches
The former problem has been fixed and the gripper now moves to all of the locations consistently. There is a 92% success rate now. Minor errors now occur. The cup would get stuck going into the press occassionally this may have been fixed with a new press. The cap occassionally slips off going into the press this is likely because the press does not hold the cap properly before the cup is released.

Also keys have been for the AWS server and should be moved to the app folder.